reduce memory used by novec trace routines

+++ src/nw_trace.c
@@ -23,12 +23,12 @@ parasail_result_t* ENAME(
         const char * const restrict _s2, const int s2Len,
         const int open, const int gap, const parasail_matrix_t *matrix)
 {
-    parasail_result_t *result = parasail_result_new_trace(s1Len, s2Len, 16, sizeof(int));
+    parasail_result_t *result = parasail_result_new_trace(s1Len, s2Len, 16, sizeof(int8_t));
     int * const restrict s1 = parasail_memalign_int(16, s1Len);
     int * const restrict s2 = parasail_memalign_int(16, s2Len);
     int * const restrict H = parasail_memalign_int(16, s2Len+1);
     int * const restrict F = parasail_memalign_int(16, s2Len+1);
-    int * const restrict HT = (int* const restrict)result->trace->trace_table;
+    int8_t * const restrict HT = (int8_t* const restrict)result->trace->trace_table;
     int i = 0;
     int j = 0;
